refactor(types): centralize transazioni types
