+++ Frontend-nextjs/types/index.ts
@@ -4,3 +4,4 @@ export * from "./types/category";
 export * from "./types/ricorrenza";
 export * from "./types/user";
 export * from "./profilo/row";
+export * from "./transazioni";
