+++ Frontend-nextjs/types/transazioni/index.ts
@@ -1,4 +1,4 @@
-// types.ts
+// Tipi specifici per la pagina Transazioni
 import { Transaction } from "@/types/types/transaction";
 
 // Estende Transaction con campo di raggruppamento
